+++ packages/browser-wallet/src/popup/pages/Settings/i18n/da.ts
@@ -6,6 +6,7 @@ const t: typeof en = {
     network: 'Netværksindstillinger',
     recover: 'Genskab ID & konti',
     about: 'Om',
+    seedPhrase: 'Vis recovery phrase',
     toggle: {
         dark: 'Skift til mørkt udseende',
         light: 'Skift til lyst udseende',