+++ packages/browser-wallet/src/popup/pages/ViewSeedPhrase/i18n/da.ts
@@ -1,0 +1,11 @@
+import type en from './en';
+
+const t: typeof en = {
+    title: 'Vis recovery phrase',
+    description: 'Indtast venligst din adgangskode for at vise dit recovery phrase.',
+    copyDescription: 'Tryk på knappen for at kopiere dit recovery phrase.',
+    show: 'Vis recovery phrase',
+    done: 'Færdig',
+};
+
+export default t;