+++ packages/browser-wallet/src/popup/pages/ViewSeedPhrase/i18n/en.ts
@@ -1,0 +1,9 @@
+const t = {
+    title: 'View recovery phrase',
+    description: 'Please enter your passcode to show the recovery phrase.',
+    copyDescription: 'Press the button to copy your recovery phrase.',
+    show: 'Show recovery phrase',
+    done: 'Done',
+};
+
+export default t;