+++ packages/browser-wallet/src/popup/shell/i18n/locales/da.ts
@@ -26,6 +26,7 @@ import addWeb3IdCredential from '@popup/pages/AddWeb3IdCredential/i18n/da';
 // TODO dansk
 import verifiableCredentialBackup from '@popup/pages/VerifiableCredentialBackup/i18n/en';
 import ageProofRequest from '@popup/pages/AgeProofRequest/i18n/da';
+import viewSeedPhrase from '@popup/pages/ViewSeedPhrase/i18n/da';
 
 import type en from './en';
 
@@ -57,6 +58,7 @@ const t: typeof en = {
     verifiableCredential,
     verifiableCredentialBackup,
     ageProofRequest,
+    viewSeedPhrase,
 };
 
 export default t;